Fetch environment from GCS

+++ poltergust/__init__.py
@@ -1,7 +1,10 @@
+import os
 import os.path
 import yaml
+from pathlib import Path
 import sys
 import traceback
+import zipfile
 import luigi
 import luigi.local_target
 import pieshell
@@ -12,6 +15,7 @@ import requests
 import datetime
 import poltergust_luigi_utils # Add GCS luigi opener
 import poltergust_luigi_utils.logging_task
+import poltergust_luigi_utils.gcs_opener
 
 DB_URL = os.environ.get("DB_URL")
 
@@ -29,7 +33,29 @@ def make_environment(envpath, environment, log):
     for dep in environment["dependencies"]:
         for line in _.pip.install(dep):
             log(line)
-        
+
+def zip_dir(envdir, log):
+    archive = envdir + ".zip"
+    envdir = Path(envdir)
+    with zipfile.ZipFile(archive, 'w', compression=zipfile.ZIP_DEFLATED, compresslevel=6) as f:
+        for file in envdir.rglob('*'):
+            f.write(file, arcname=file.relative_to(envdir))
+        log("Zipped: {}".format(f))
+    return archive
+
+
+def download_environment(envpath, path, log):
+    if not os.path.exists(envpath):
+        envdir = os.path.dirname(envpath)
+        if not os.path.exists(envdir):
+            os.makedirs(envdir)
+
+    with poltergust_luigi_utils.client.download(path) as z:
+        with zipfile.ZipFile(z, mode="r") as arc:
+            arc.extractall(envpath)
+        log(arc.printdir())
+
+
 class MakeEnvironment(poltergust_luigi_utils.logging_task.LoggingTask, luigi.Task):
     path = luigi.Parameter()
     hostname = luigi.Parameter()
@@ -37,16 +63,24 @@ class MakeEnvironment(poltergust_luigi_utils.logging_task.LoggingTask, luigi.Tas
 
     def run(self):
         with self.logging(self.retry_on_error):
-            with luigi.contrib.opener.OpenerTarget(self.path).open("r") as f:
-                environment = yaml.load(f, Loader=yaml.SafeLoader)
-            make_environment(self.envdir().path, environment, self.log)
-            with self.output().open("w") as f:
-                f.write("DONE")        
-        
+            zip_path = str(self.path) + ".zip"
+            # if poltergust_luigi_utils.gcs_opener.client.exists(zip_path):
+            if luigi.contrib.opener.OpenerTarget(zip_path).exists():
+                download_environment(self.envdir().path, zip_path, self.log)
+            else:
+                with luigi.contrib.opener.OpenerTarget(self.path).open("r") as f:
+                    environment = yaml.load(f, Loader=yaml.SafeLoader)
+                    make_environment(self.envdir().path, environment, self.log)
+                    with self.output().open("w") as f:
+                        f.write("DONE")        
+                archived_env = zip_dir(self.envdir().path, self.log)
+                poltergust_luigi_utils.gcs_opener.client.put(archived_env, zip_path)
+
+
     def envdir(self):
         return luigi.local_target.LocalTarget(
             os.path.join("/tmp/environments", self.path.replace("://", "/").lstrip("/")))
-    
+
     def logfile(self):
         return luigi.contrib.opener.OpenerTarget("%s.%s.log.txt" % (self.path, self.hostname))
